Refactor integration tests module
