+++ spring-pulsar-test/src/test/kotlin/com/intuit/pulsar/tests/consumers/SharedConsumer.kt
@@ -1,7 +1,7 @@
 package com.intuit.pulsar.tests.consumers
 
 
-import com.intuit.pulsar.tests.TestConstants
+import com.intuit.pulsar.tests.utils.TestConstants
 import com.intuit.spring.pulsar.client.annotations.consumer.PulsarConsumer
 import com.intuit.spring.pulsar.client.annotations.consumer.Subscription
 import com.intuit.spring.pulsar.client.annotations.consumer.Topic
